pyproject.toml
[project]
name = "bakalari-sumary"
version = "0.1.0"
description = "Create static HTML output from bakalari API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]
